requirements.txt
streamlit>=1.32.0
spacy>=3.7.0
openai>=1.30.0
numpy>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
